Foi adicionado o botão sair ao sidebar, foi transformado em modal a opção de delete e também foi formatado os tamanhos
