+++ src/views/LoginView.vue
@@ -1,5 +1,5 @@
 <script setup>
-import LoginComponent from '../components/LoginComponent.vue';
+import LoginComponent from '../components/UsuariosComponents/Login/LoginComponent.vue';
 import FooterComponent from '../components/FooterComponent.vue';
 import NavBarComponent from '../components/NavBarComponent.vue'; 
 </script>
