+++ src/views/admin/BACKUP-TRASH/DeleteProductView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import DeleteProductsComponents from '@/components/AdminComponents/TRASH-BACKUP/DeleteProductsComponents.vue';
+</script>
+<template>
+<div>
+  <DeleteProductsComponents/> 
+</div>
+</template>
+
