+++ src/views/admin/BACKUP-TRASH/EditProductView.vue
@@ -1,0 +1,9 @@
+<script setup>
+import EditProductsComponents from '@/components/AdminComponents/TRASH-BACKUP/EditProductsComponents.vue';
+</script>
+<template>
+<div>
+  <EditProductsComponents/> 
+</div>
+</template>
+
