+++ src/views/admin/BACKUP-TRASH/RegisterProductView.vue
@@ -1,0 +1,24 @@
+<script setup>
+  // import Card from '@/components/AdminComponents/CadastrarProdutosComponent.vue';
+
+</script>
+
+<template>
+  <div>
+    <Card />
+  </div>
+</template>
+  
+  
+  
+<style>
+#app {
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  color: #2c3e50;
+
+}
+</style>
+
