--- src/views/admin/DeleteProductView.vue
@@ -1,9 +1,0 @@
-<script setup>
-import DeleteProductsComponents from '@/components/AdminComponents/DeleteProductsComponents.vue';
-</script>
-<template>
-<div>
-  <DeleteProductsComponents/> 
-</div>
-</template>
-
