--- src/views/admin/EditProductView.vue
@@ -1,9 +1,0 @@
-<script setup>
-import EditProductsComponents from '@/components/AdminComponents/EditProductsComponents.vue';
-</script>
-<template>
-<div>
-  <EditProductsComponents/> 
-</div>
-</template>
-
