--- src/views/admin/RegisterProductView.vue
@@ -1,24 +1,0 @@
-<script setup>
-  // import Card from '@/components/AdminComponents/CadastrarProdutosComponent.vue';
-
-</script>
-
-<template>
-  <div>
-    <Card />
-  </div>
-</template>
-  
-  
-  
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-
-}
-</style>
-
